flask-server/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
Flask-MongoEngine==1.0.0
python-dotenv==1.0.0
